+++ main/parsers_p.h
@@ -173,6 +173,7 @@
 	TclParser, \
 	TclOOParser, \
 	TerraformParser, \
+	TerraformVariablesParser, \
 	TexParser, \
 	TexBeamerParser, \
 	TTCNParser, \